re-organizing code & renaming files. general tidy-up
